+++ frontend/src/styles/pages/_training-page.scss
@@ -29,6 +29,17 @@
         }
       }
 
+      .form-overlay {
+        background-color: $black;
+        opacity: 0.5;
+        position: fixed;
+        top: 0;
+        left: 0;
+        width: 100%;
+        height: 100%;
+        z-index: 99;
+      }
+
       .usa-search {
         max-width: none;
         width: 100%;
@@ -38,7 +49,19 @@
         }
 
         @media screen and (max-width: ($tablet-md - 1)) {
-          display: none;
+          z-index: 100;
+          background-color: $secondary-extra-light;
+          flex-wrap: wrap;
+          padding: $size-24 $size-16 $size-32;
+          position: fixed;
+          bottom: 0;
+          left: 0;
+        }
+
+        label {
+          font-size: $size-24;
+          font-weight: 700;
+          margin-bottom: $size-8;
         }
 
         input {
